requirements.txt
Django==3.1
psycopg2-binary==2.9.1
Pillow==8.3.1
djangorestframework==3.12.2
drf-yasg